Update openCard parameters
