+++ src/internal/mediaUtil.ts
@@ -1,4 +1,3 @@
-import { people } from '../public/people';
 import { media } from '../public/media';
 
 /**
@@ -99,37 +98,6 @@ export function validateScanBarCodeInput(barCodeConfig: media.BarCodeConfig): bo
       barCodeConfig.timeOutIntervalInSec > 60
     ) {
       return false;
-    }
-  }
-  return true;
-}
-
-/**
- * Returns true if the people picker params are valid and false otherwise
- */
-export function validatePeoplePickerInput(peoplePickerInputs: people.PeoplePickerInputs): boolean {
-  if (peoplePickerInputs) {
-    if (peoplePickerInputs.title) {
-      if (typeof peoplePickerInputs.title !== 'string') {
-        return false;
-      }
-    }
-
-    if (peoplePickerInputs.setSelected) {
-      if (typeof peoplePickerInputs.setSelected !== 'object') {
-        return false;
-      }
-    }
-
-    if (peoplePickerInputs.openOrgWideSearchInChatOrChannel) {
-      if (typeof peoplePickerInputs.openOrgWideSearchInChatOrChannel !== 'boolean') {
-        return false;
-      }
-    }
-    if (peoplePickerInputs.singleSelect) {
-      if (typeof peoplePickerInputs.singleSelect !== 'boolean') {
-        return false;
-      }
     }
   }
   return true;
